backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
databases[postgresql]
pandas
geopandas
shapely
psycopg2-binary
python-dotenv
pydantic-settings
orjson
